force int for item weights

+++ app/cli.py
@@ -1,4 +1,5 @@
 import argparse
+
 from models.cli import Cli
 
 
@@ -6,6 +7,6 @@ def cli():
     parser = argparse.ArgumentParser()
     parser.add_argument("--input", required=True, help="Path to input CSV file")
     parser.add_argument("--output", required=True, help="Path to output CSV file")
-    parser.add_argument("--sum", required=True, type=float, help="Target sum")
+    parser.add_argument("--sum", required=True, type=int, help="Target sum")
     args = parser.parse_args()
     return Cli(sum=args.sum, input=args.input, output=args.output)
