+++ models/binpacking.py
@@ -3,4 +3,4 @@ from pydantic import BaseModel
 
 class Item(BaseModel):
     name: str
-    weight: float
+    weight: int
